+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [Unreleased]
 
+### Added
+
+- Add optional delimiter grouping to `list`, with distinguishable file and virtual-prefix outputs.
+
 ## [1.2.0] - 2026-08-04
 
 ### Security